+++ _Typings/gameApi/modules/variables.d.ts
@@ -1,28 +1,17 @@
 declare module "game:module" {
-    type builtInCollections =
-        | "Score"
-        | "Cure"
-        | "Recover"
-        | "Froze"
-        | "Electric"
-        | "Key_Red"
-        | "Key_Blue"
-        | "Key_Green"
-        | "Key_Cyan"
-        | "Key_Magenta"
-        | "Key_Yellow"
+    import { BuiltinVariables } from "game:type"
 
     namespace variables {
         /**
          * @param name the name of the variable to get, if starts with `C_` it refers to a `Collection`
-         * @builtinCollections Click to see {@link builtInCollections}
+         * @BuiltinVariables click to see {@link BuiltinVariables | all built-in variables}
          */
         const get: <T extends string>(name: T) => T extends `C_${string}` ? int : any
         /**
          * @param name the name of the variable to set, if starts with `C_` it refers to a `Collection`
          * @param value the value to set. `Collection` values should be `int`
          * @NOTE **NEVER** name a variable starting with `C_` manually
-         * @builtinCollections Click to see {@link builtInCollections}
+         * @BuiltinVariables click to see {@link BuiltinVariables | all built-in variables}
          */
         const set: <T extends string>(name: T, value: T extends `C_${string}` ? int : any) => void
         /**