+++ _Typings/gameApi/types/builtin.d.ts
@@ -1,0 +1,16 @@
+declare module "game:type" {
+    type BuiltinCollections =
+        | "Score"
+        | "Cure"
+        | "Recover"
+        | "Froze"
+        | "Electric"
+        | "Key_Red"
+        | "Key_Blue"
+        | "Key_Green"
+        | "Key_Cyan"
+        | "Key_Magenta"
+        | "Key_Yellow"
+
+    type BuiltinVariables = "C_Key_Red" | "C_Key_Blue" | "C_Key_Green" | "C_Key_Cyan" | "C_Key_Magenta" | "C_Key_Yellow"
+}