+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
+## v1.1.2 (2025-07-02)
+
+-   Renamed type `builtInCollections` to `BuiltinCollections`, and moved it to a new file `builtin.d.ts`, where another type `BuiltinVariables` is defined.
+-   Moved type `BallType` to `player.d.ts`.
+-   Fixed a bug that the type of `Events.OnPreGetCollection` and `Events.OnPostGetCollection` should be `string[]` instead of `` `C_${string}`[] ``.
+-   Fixed some wrong descriptions.
+-   Added `dialogWindowManager.openMessageDialog()` and `dialogWindowManager.openSubWindow()`.
+-   Changed the type of `editor.version` from `int` to `string`.
+-   Added some badges in `README.md`.
+
 ## v1.1.1 (2025-06-30)
 
 -   Added `levelManager.spawnVfxPRS()`.